+++ src/lmgrep/lucene/dictionary.clj
@@ -1,6 +1,5 @@
 (ns lmgrep.lucene.dictionary
-  (:require [clojure.core.reducers :as r]
-            [jsonista.core :as json]
+  (:require [jsonista.core :as json]
             [lucene.custom.query :as q]
             [lmgrep.lucene.analyzer :as analyzer]
             [lmgrep.lucene.field-name :as field-name])
@@ -9,34 +8,6 @@
            (org.apache.lucene.search Query)
            (org.apache.lucene.analysis Analyzer)
            (clojure.lang PersistentArrayMap)))
-
-(defn prepare-metadata
-  "Metadata must be a map String->String"
-  [type meta]
-  (let [str->str (if meta
-                   (reduce-kv (fn [m k v] (assoc m (name k) v)) {} meta)
-                   {})]
-    (assoc str->str "_type" type)))
-
-(defn query->monitor-query [questionnaire-entry field-name monitor-analyzer]
-  (try
-    (let [query (get questionnaire-entry :query)
-          query-parser-name (keyword (get questionnaire-entry :query-parser))
-          query-parser-conf (get questionnaire-entry :query-parser-conf)]
-      (MonitorQuery. ^String (get questionnaire-entry :id)
-                     ^Query (q/parse query query-parser-name query-parser-conf field-name monitor-analyzer)
-                     ^String query
-                     (prepare-metadata (get questionnaire-entry :type) (get questionnaire-entry :meta))))
-    (catch ParseException e
-      (when (System/getenv "DEBUG_MODE")
-        (.println System/err (format "Failed to parse query: '%s' with exception '%s'" questionnaire-entry e))
-        (.printStackTrace e))
-      (throw e))
-    (catch Exception e
-      (when (System/getenv "DEBUG_MODE")
-        (.println System/err (format "Failed create query: '%s' with '%s'" questionnaire-entry e))
-        (.printStackTrace e))
-      (throw e))))
 
 (defrecord Dict [field-name monitor-query analysis-conf])
 
@@ -80,16 +51,40 @@
           (.printStackTrace e))
         (throw e)))))
 
+(defn in-memory-query-constructor [custom-analyzers]
+  (fn monitor-query
+    [{:keys [id query meta default-field-name query-parser-name query-parser-conf analyzer] :as mq}]
+    (try
+      (let [monitor-analyzer (get-string-analyzer analyzer custom-analyzers)]
+        (MonitorQuery. ^String id
+                       ^Query (q/parse query query-parser-name query-parser-conf default-field-name monitor-analyzer)
+                       ^String query
+                       (prepare-meta meta)))
+      (catch ParseException e
+        (when (System/getenv "DEBUG_MODE")
+          (.println System/err (format "Failed to parse query: '%s' with exception '%s'" mq e))
+          (.printStackTrace e))
+        (throw e))
+      (catch Exception e
+        (when (System/getenv "DEBUG_MODE")
+          (.println System/err (format "Failed create query: '%s' with '%s'" mq e))
+          (.printStackTrace e))
+        (throw e)))))
+
+(defn stable-id [m]
+  (str (Math/abs ^int (.hashCode ^PersistentArrayMap m))))
+
 (defn prepare-query-entry
-  [questionnaire-entry default-type global-analysis-conf custom-analyzers monitor-query-constructor-fn]
-  (let [analysis-conf (if (empty? (get questionnaire-entry :analysis))
+  [questionnaire-entry default-type global-analysis-conf monitor-query-constructor-fn]
+  (let [id (or (:id questionnaire-entry) (stable-id questionnaire-entry))
+        analysis-conf (if (empty? (get questionnaire-entry :analysis))
                         global-analysis-conf
                         (assoc (get questionnaire-entry :analysis)
                           :config-dir (get global-analysis-conf :config-dir)))
         ; parameter for the query parser
         default-field-name (get-field-name analysis-conf)
         monitor-query (monitor-query-constructor-fn
-                        {:id                 (get questionnaire-entry :id)
+                        {:id                 id
                          :query              (get questionnaire-entry :query)
                          :meta               (assoc (get questionnaire-entry :meta) "_type" default-type)
                          :default-field-name default-field-name
@@ -120,24 +115,15 @@
   [questionnaire default-type options custom-analyzers]
   (let [global-analysis-conf (assoc (get options :analysis)
                                :config-dir (get options :config-dir))
-        monitor-query-constructor-fn (monitor-query-constructor custom-analyzers)]
-    (->> questionnaire
-         (r/map (fn [questionnaire-entry]
-                  (if (get questionnaire-entry :id)
-                    questionnaire-entry
-                    (assoc questionnaire-entry :id (str (Math/abs ^int (.hashCode ^PersistentArrayMap questionnaire-entry)))))))
-         (r/map (fn [questionnaire-entry]
-                  (prepare-query-entry (handle-query-parser-settings questionnaire-entry options)
-                                       default-type
-                                       global-analysis-conf
-                                       custom-analyzers
-                                       monitor-query-constructor-fn)))
-         (r/foldcat))))
+        monitor-query-constructor-fn (if (:queries-index-dir options)
+                                       (monitor-query-constructor custom-analyzers)
+                                       (in-memory-query-constructor custom-analyzers))]
+    (mapv (fn [questionnaire-entry]
+            (prepare-query-entry (handle-query-parser-settings questionnaire-entry options)
+                                 default-type
+                                 global-analysis-conf
+                                 monitor-query-constructor-fn)) questionnaire)))
 
 (defn get-monitor-queries
   "Returns a vector of MonitorQuery"
-  [questionnaire]
-  (->> questionnaire
-       (r/map :monitor-query)
-       (r/remove nil?)
-       (into [])))
+  [questionnaire] (mapv :monitor-query questionnaire))